[ty] Move binary expression logic out of `builder.rs`
